Move getFloatingIps functions to hcloud.go
